Revert "Fix: Only call setOriginalImageSize() on single page images (#394)"
This reverts commit 8f1cc35be9de6f2608d6934522504be53365c86b.
